+++ kafka/error.go
@@ -22,7 +22,7 @@ package kafka
 
 /*
 #include <stdlib.h>
-#include "rdkafka_select.h"
+#include "select_rdkafka.h"
 */
 import "C"